ml/requirements.txt
fastapi==0.115.0
uvicorn==0.30.6
scikit-learn==1.5.2
pandas==2.2.2
joblib==1.4.2
numpy==2.0.2
pillow==10.4.0
requests==2.32.3
